Fix add invoice: switch to service role key + fix amount parsing
Invoice API was using anon key which may lack INSERT permissions.
Switched to service role key. Also fixed parseInt → parseFloat for
decimal amounts.

+++ src/app/business/finances/page.tsx
@@ -288,7 +288,7 @@ export default function Finances() {
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify({
           gig_title: newInvoice.gig_title,
-          amount: parseInt(newInvoice.amount),
+          amount: parseFloat(newInvoice.amount),
           currency: newInvoice.currency,
           type: newInvoice.type,
           due_date: newInvoice.due_date,
